Chapter 3 - Multiple Bucket Registries
This commit uses multiple Bucket Registries as to allow concurrent requests to
performed at the same time against buckets that live in different registries.

This removes the bottleneck of having all requests pass through the same
BucketRegistry as in Elixir, a process can only handle one message at a time.